Представления для списка всех пользователей (с сортировкой) вынесены в отдельный пакет
