+++ recipeblog/urls.py
@@ -3,7 +3,7 @@ from . import views
 
 urlpatterns = [
     path('', views.index, name='home'),
-    path('users/', include('users.urls')),
+    path('users/', include('packages.users.urls')),
     path('users/<int:pk>', views.show_user_profile, name='user-detail'),
     path('usersfavs/<int:pk>/', include('usersfavs.urls')),
     path('usersrecipes/<int:pk>/', include('usersrecipes.urls')),
